% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective_init}
\alias{objective_init}
\title{Objective Function for Init Object}
\usage{
objective_init(init_object, lambda, weights)
}
\arguments{
\item{init_object}{List object containing matrices}

\item{lambda}{Scalar for spatial smoothing}

\item{weights}{Matrix with weights}
}
\value{
Returns the value of the objective function.
}
\description{
Calculates the objective function based on the matrices stored in the init object.
}
\keyword{internal}
