% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objective_matrix}
\alias{objective_matrix}
\title{Calculation of Objective Function}
\usage{
objective_matrix(matrix_list, lambda, weights)
}
\arguments{
\item{matrix_list}{A list of matrices \eqn{K_i}.}

\item{lambda}{Scalar smoothing parameter.}

\item{weights}{Matrix of weights.}
}
\value{
Returns the value of the objective function.
}
\description{
Calculation of the value of the objective function for a given list of matrices,
lambda, and a weighting matrix.
}
\keyword{internal}
