% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRPCAMS_methods.R
\name{summary.msPCA}
\alias{summary.msPCA}
\title{Summary Method for msPCA Objects}
\usage{
\method{summary}{msPCA}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"msPCA"} as returned by the main msPCA fitting function.}

\item{...}{Currently ignored.}
}
\value{
Prints a summary to the console. No return value.
}
\description{
Provides a summary of a sparse multi-source PCA (`msPCA`) result, including group-wise sparsity
and explained variance.
}
