% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssMRCD_methods.R
\name{summary.ssMRCD}
\alias{summary.ssMRCD}
\title{Summary Method for ssMRCD Object}
\usage{
\method{summary}{ssMRCD}(object, ...)
}
\arguments{
\item{object}{object of class \code{"ssMRCD"}, output of \code{\link[ssMRCD]{ssMRCD}}.}

\item{...}{further parameters.}
}
\value{
Prints a summary of the \code{ssMRCD} object.
}
\description{
Summarises most important information of output \code{\link[ssMRCD]{ssMRCD}}.
}
\seealso{
See also \code{\link[ssMRCD]{ssMRCD}}, \code{\link[ssMRCD]{plot.ssMRCD}}.
}
