% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helpers.R
\name{time_weights}
\alias{time_weights}
\title{Band weight matrix for time series groupings}
\usage{
time_weights(N, off_diag)
}
\arguments{
\item{N}{number of groups.}

\item{off_diag}{vector for off-diagonal values unequal to zero.}
}
\value{
Returns weight matrix for time series groups appropriate for \code{\link[ssMRCD]{ssMRCD}}.
}
\description{
Band weight matrix for time series groupings
}
\examples{
time_weights(N = 10, off_diag = c(2,1))

}
\seealso{
\code{\link[ssMRCD]{geo_weights}}, \code{\link[ssMRCD]{rescale_weights}}
}
